To better organize data and the frontend project, move all data to a folder called data on the src directory
