+++ src/data/medicalConditionsList.jsx
